Patch 3 to Patch 2 (#1)
* Create make_accuracy_plot.py

make_accuracy_plot.py is how we make the accuracy plot from for the database, this version of the code includes the function: "make_fake_plot" which is needed as a placeholder for the acc_bytes in processs_im.py. the placeholder is needed because the plot can not be completed until the connection to Postgres has been established. Once the Connection is established, PM offsets are read from the database and the Accuracy Plot is completed.

* Update process_im.py

Needed to Import New Functions

* process_im.py

added grab Sensor Offsets

* Update upload_inspect.py

Added GrabSensorOffsets Function For Use In process_im.py